refactor: move the Block, CellDep, and Transaction types to src/models
