+++ src/components/TransactionItem/TransactionItemCellList/index.tsx
@@ -3,6 +3,7 @@ import { Link } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
 import TransactionCellListPanel from './styled'
 import { Cell } from '../../../models/Cell'
+import { Transaction } from '../../../models/Transaction'
 
 const MAX_CELL_SHOW_SIZE = 10
 
@@ -12,7 +13,7 @@ export default ({
   render,
 }: {
   cells: Cell[]
-  transaction: State.Transaction
+  transaction: Transaction
   render: (cell: Cell) => ReactNode
 }) => {
   const { t } = useTranslation()
